Adjusted to use SOCKET and size_t types where applicable.

+++ unisim/unisim/unisim.cpp
@@ -230,12 +230,12 @@ void beam_multi_collision()
 
 void print_positions()
 {
-    for (uint32_t i = 0 ; i < objs.size() ; i++)
+    for (size_t i = 0 ; i < objs.size() ; i++)
     {
         fprintf(stderr, "PO%u   %g   %g   %g\n", i, objs[i]->position.x, objs[i]->position.y, objs[i]->position.z);
     }
 
-    for (uint32_t i = 0 ; i < beams.size() ; i++)
+    for (size_t i = 0 ; i < beams.size() ; i++)
     {
         fprintf(stderr, "BM%u   %g   %g   %g\n", i, beams[i]->front_position.x, beams[i]->front_position.y, beams[i]->front_position.z);
     }
